feat: provide es6 module
