+++ src/ChunkProcessorInterceptor.js
@@ -1,20 +1,27 @@
 /* jslint node: true, esnext: true */
 'use strict';
 
-const Interceptor = require('kronos-interceptor').Interceptor;
-const parserFactory = require('./data-processor-chunk');
+import {
+	Interceptor
+} 
+from 'kronos-interceptor';
+
+import {
+	DataProcessorChunkFactory
+}
+from './data-processor-chunk';
 
 /**
  * This interceptor cares about the handling of the messages.
  * It will add the hops and copies the messages
  */
-class ChunkProcessorInterceptor extends Interceptor {
+export default class ChunkProcessorInterceptor extends Interceptor {
 
 	constructor(config, endpoint) {
 		super(config, endpoint);
 
 		// just validate the config once
-		parserFactory(config.config, true);
+		DataProcessorChunkFactory(config.config, true);
 	}
 
 	static get name() {
@@ -23,11 +30,10 @@ class ChunkProcessorInterceptor extends Interceptor {
 
 	receive(request, oldRequest) {
 		if (request.payload) {
-			const streamFilter = parserFactory(this.config.config);
+			const streamFilter = DataProcessorChunkFactory(this.config.config);
 			const stream = request.payload;
 			request.payload = stream.pipe(streamFilter);
 		}
 		return this.connected.receive(request, oldRequest);
 	}
 }
-exports.Interceptor = ChunkProcessorInterceptor;
