+++ tests/interceptor_test.js
@@ -11,24 +11,24 @@ const chai = require('chai'),
   assert = chai.assert,
   expect = chai.expect,
   should = chai.should(),
-  InterceptorUnderTest = require('../index').Interceptor,
+  InterceptorUnderTest = require('../dist/module').ChunkProcessorInterceptor,
   MockReceiveInterceptor = require('kronos-test-interceptor').MockReceiveInterceptor;
 
 
 const stepMock = {
-  "name": "dummy step name",
-  "type": "dummy step type"
+  name: 'dummy step name',
+  type: 'dummy step type'
 };
 
 const checkProperties = {
-  "config": {
-    "keyHashFields": ["first", "last"],
-    "keyHashName": "__key",
-    "contentHashFields": ["first", "last", "friends"],
-    "contentHashName": "__content",
-    "scopeHashFields": ["street"],
-    "scopeHashName": "__scope",
-    "multiRowFields": ["friends"]
+  config: {
+    keyHashFields: ["first", "last"],
+    keyHashName: "__key",
+    contentHashFields: ["first", "last", "friends"],
+    contentHashName: "__content",
+    scopeHashFields: ["street"],
+    scopeHashName: "__scope",
+    multiRowFields: ["friends"]
   }
 };
 
@@ -37,8 +37,8 @@ describe('Interceptor test', function () {
 
   it('Create', function () {
     const endpoint = {
-      "owner": stepMock,
-      "name": "gumboIn"
+      owner: stepMock,
+      name: "gumboIn"
     };
     const messageHandler = new InterceptorUnderTest(checkProperties, endpoint);
     assert.ok(messageHandler);
@@ -46,12 +46,12 @@ describe('Interceptor test', function () {
 
   it('Send message', function (done) {
     const endpoint = {
-      "owner": stepMock,
-      "name": "gumboIn"
+      owner: stepMock,
+      name: "gumboIn"
     };
 
     const sendMessage = {
-      "info": "first message"
+      info: "first message"
     };
 
     const messageHandler = new InterceptorUnderTest(checkProperties, endpoint);
@@ -60,7 +60,7 @@ describe('Interceptor test', function () {
       assert.ok(request);
 
       assert.deepEqual(request, {
-        "info": "first message"
+        info: "first message"
       });
       done();
     });
@@ -70,7 +70,5 @@ describe('Interceptor test', function () {
     messageHandler.receive(sendMessage);
 
   });
-
-
 
 });
